Refactor workspace commands to move logic to core.
